% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky_diagonalcpp}
\alias{cholesky_diagonalcpp}
\title{Cholesky diagonal update}
\arguments{
\item{R}{n by n matrix, right triangular Cholesky factor R, where t(R)*R = A}

\item{diagadd}{n by 1 vector, diagonal adding vector}
}
\description{
It calculates the Cholesky factor of (A + diag(diagadd)) directly from chol(A) and diagadd vector. See Algorithm 2 in Appendix A1 of Lee et al.(2021)
}
\examples{

X = matrix(rnorm(1000*200),1000, 200)
A = crossprod(X) # positive definite A
R = chol(A) 
diagadd = runif(200)

Rnew = cholesky_diagonalcpp(R, diagadd) 
all.equal(Rnew, chol(crossprod(X)+diag(diagadd)))

}
\references{
Lee, C. J., Luo, Z. T., & Sang, H. (2021). T-LoHo: A Bayesian Regularization Model for Structured Sparsity and Smoothness on Graphs. \emph{Advances in Neural Information Processing Systems 34 (NeurIPS 2021)}.
}
