% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tloho_lm_normalmeans.R
\name{tloho_lm_normalmeans}
\alias{tloho_lm_normalmeans}
\title{TLoHolm: Tree-based Low rank Horseshoe regularization prior in linear model}
\usage{
tloho_lm_normalmeans(
  Y,
  graph0,
  init_val = NULL,
  c = 0.5,
  tau0 = 1,
  MCMC = 50000,
  BURNIN = 40000,
  THIN = 10,
  Dahl = F,
  seed = NULL
)
}
\arguments{
\item{Y}{n by 1 scalar, real-valued response variables.}

\item{graph0}{igraph object, reflecting the structure of beta.}

\item{init_val}{list with name 'beta' and/or 'trees', Initial value of beta and/or spanning forest.}

\item{c}{model size penalization hyperparameter, value between 0 and 1. c = 1}

\item{tau0}{shrinkage strength hyperparameter, positive real.}

\item{MCMC}{number of total MCMC iteration.}

\item{BURNIN}{number of burn-in iteration which will be discarded}

\item{THIN}{thin-in rate. The final number of posterior sample is nsamples = (MCMC-BURNIN)/THIN.}

\item{Dahl}{bool, whether to calculate Dahl's cluster estimate after MCMC. Note that when p is large, this can take longer than MCMC itself}

\item{seed}{seed.}

\item{X}{n by p matrix, real-valued predictors. If not standardized, it will be standardized so that all columns have L2 norm 1}
}
\value{
A list containing:\tabular{ll}{
   \code{beta_out} \tab nsamples by p matrix, posterior samples of beta\cr
   \tab \cr
   \code{lambda2_out} \tab list of length nsamples, posterior samples of lambda^2 \cr
   \tab \cr
   \code{tau2_out} \tab length nsamples vector, posterior samples of tau^2 \cr
   \tab \cr
   \code{sigmasq_y_out} \tab length nsamples vector, posterior samples of sigma^2 \cr
   \tab \cr
   \code{cluster_out} \tab nsamples by p matrix, posterior samples of cluster label \cr
   \tab \cr
   \code{cluster_map} \tab length p vector, maximum a posteriori cluster estimate \cr
   \tab \cr
   \code{cluster_est_Dahl} \tab (if Dahl = True)length p vector, cluster estimate based on Dahl's method  \cr
   \tab \cr
   \code{log_post_out} \tab length nsamples vector, log posterior likelihood(up to a constant) \cr
   \tab \cr
   \code{acc} \tab list with split/merge/change/hyper move proposal counts and acceptance counts \cr
   \tab \cr
   \code{map_beta_est} \tab length p vector, maximum a posteriori beta estimate\cr
   \tab \cr
   \code{mean_beta_est} \tab length p vector, posterior mean beta estimate \cr
   \tab \cr
   \code{median_beta_est} \tab length p vector, posterior median beta estimate \cr
   \tab \cr
   \code{map_MSF_est} \tab igraph object, a sample of minimum spanning forest when posterior likelihood is maximized\cr
   \tab \cr
}
}
\description{
TLoHolm: Tree-based Low rank Horseshoe regularization prior in linear model
}
\details{
Anonymous authors, 
Title: T-LoHo: A Bayesian Regularization Model for Structured Sparsity and Smoothness on Graphs
submitted to NeurIPS2021, paper2212
}
