% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tloho_lm.R
\name{tloho_lm}
\alias{tloho_lm}
\title{tloho_lm: Tree-based Low rank Horseshoe regularization prior in linear model}
\usage{
tloho_lm(
  Y,
  X,
  graph0,
  init_val = NULL,
  c = 0.5,
  tau0 = 1,
  MCMC = 50000,
  BURNIN = 40000,
  THIN = 10,
  loss = "binder",
  hsplus = F,
  seed = NULL
)
}
\arguments{
\item{Y}{n by 1 scalar, real-valued response variables.}

\item{X}{n by p matrix, real-valued predictors. If not standardized, it will be standardized so that all columns have L2 norm 1}

\item{graph0}{igraph object, reflecting the structure of beta.}

\item{init_val}{list with name 'beta' and/or 'trees', Initial value of beta and/or spanning forest.}

\item{c}{model size penalization hyperparameter, value between 0 and 1. c = 1}

\item{tau0}{shrinkage strength hyperparameter, positive real.}

\item{MCMC}{number of total MCMC iteration.}

\item{BURNIN}{number of burn-in iteration which will be discarded}

\item{THIN}{thin-in rate. The final number of posterior sample is nsamples = (MCMC-BURNIN)/THIN.}

\item{loss}{"binder" or "VI", whether to calculate cluster estimate based on Binder loss (default) or VI loss}

\item{hsplus}{default F, use horseshoe+ prior instead of horseshoe? (experimental)}

\item{seed}{seed.}
}
\value{
A list containing:\tabular{ll}{
   \code{beta_out} \tab nsamples by p matrix, posterior samples of beta\cr
   \tab \cr
   \code{lambda2_out} \tab list of length nsamples, posterior samples of lambda^2 \cr
   \tab \cr
   \code{tau2_out} \tab length nsamples vector, posterior samples of tau^2 \cr
   \tab \cr
   \code{sigmasq_y_out} \tab length nsamples vector, posterior samples of sigma^2 \cr
   \tab \cr
   \code{cluster_out} \tab nsamples by p matrix, posterior samples of cluster label \cr
   \tab \cr
   \code{cluster_map} \tab length p vector, maximum a posteriori cluster estimate \cr
   \tab \cr
   \code{cluster_est} \tab length p vector, cluster estimate based on Bayes estimator that minimizes expected loss \cr
   \tab \cr
   \code{log_post_out} \tab length nsamples vector, log posterior likelihood(up to a constant) \cr
   \tab \cr
   \code{acc} \tab list with split/merge/change/hyper move proposal counts and acceptance counts \cr
   \tab \cr
   \code{map_beta_est} \tab length p vector, maximum a posteriori beta estimate\cr
   \tab \cr
   \code{mean_beta_est} \tab length p vector, posterior mean beta estimate \cr
   \tab \cr
   \code{median_beta_est} \tab length p vector, posterior median beta estimate \cr
   \tab \cr
   \code{map_MSF_est} \tab igraph object, a sample of minimum spanning forest when posterior likelihood is maximized\cr
   \tab \cr
}
}
\description{
tloho_lm: Tree-based Low rank Horseshoe regularization prior in linear model
}
\details{
The main R function of the paper 

T-LoHo: A Bayesian Regularization Model for Structured Sparsity and Smoothness on Graphs
by C. J. Lee, Z. T. Luo, and H. Sang, NeurIPS 2021
}
